# auth/Crypto: refactor to use OpenSSL's EVP API