OpenSSL's EVP API encapsulates different encryption mechanisms and engines, including AES-NI, ARM NEON, VIA Padlock and possibly other hardware crypto accelerators. Considering that AES encryption optimized with AES-NI alone is around 6x faster than directly-callable implementation, there's no reason to not use it.

Even with the need to create (malloc) and free EVP contexts for each encryption, anything shows the performance improvements, for example unittests:
Before:
```
[==========] Running 8 tests from 1 test case.           
[----------] Global test environment set-up.             
[----------] 8 tests from AES                            
[ RUN      ] AES.ValidateLegacy                          
[       OK ] AES.ValidateLegacy (0 ms)                   
[ RUN      ] AES.ValidateSecret                          
[       OK ] AES.ValidateSecret (0 ms)                   
[ RUN      ] AES.Encrypt                                 
[       OK ] AES.Encrypt (0 ms)                          
[ RUN      ] AES.EncryptNoBl                             
[       OK ] AES.EncryptNoBl (0 ms)                      
[ RUN      ] AES.Decrypt                                 
[       OK ] AES.Decrypt (0 ms)                          
[ RUN      ] AES.DecryptNoBl                             
[       OK ] AES.DecryptNoBl (0 ms)                      
[ RUN      ] AES.Loop                                    
[       OK ] AES.Loop (94 ms)                            
[ RUN      ] AES.LoopKey                                 
100000 encoded in 0.156125                               
[       OK ] AES.LoopKey (156 ms)                        
[----------] 8 tests from AES (250 ms total)             
                                                         
[----------] Global test environment tear-down           
[==========] 8 tests from 1 test case ran. (250 ms total)
[  PASSED  ] 8 tests.
```
after:
```
[==========] Running 8 tests from 1 test case.           
[----------] Global test environment set-up.             
[----------] 8 tests from AES                            
[ RUN      ] AES.ValidateLegacy                          
[       OK ] AES.ValidateLegacy (0 ms)                   
[ RUN      ] AES.ValidateSecret                          
[       OK ] AES.ValidateSecret (0 ms)                   
[ RUN      ] AES.Encrypt                                 
[       OK ] AES.Encrypt (0 ms)                          
[ RUN      ] AES.EncryptNoBl                             
[       OK ] AES.EncryptNoBl (0 ms)                      
[ RUN      ] AES.Decrypt                                 
[       OK ] AES.Decrypt (1 ms)                          
[ RUN      ] AES.DecryptNoBl                             
[       OK ] AES.DecryptNoBl (0 ms)                      
[ RUN      ] AES.Loop                                    
[       OK ] AES.Loop (32 ms)                            
[ RUN      ] AES.LoopKey                                 
100000 encoded in 0.072338                               
[       OK ] AES.LoopKey (72 ms)                         
[----------] 8 tests from AES (105 ms total)             
                                                         
[----------] Global test environment tear-down           
[==========] 8 tests from 1 test case ran. (105 ms total)
[  PASSED  ] 8 tests.                                    

```
After with AES_NI disabled (``OPENSSL_ia32cap="~0x200000200000000" bin/unittest_crypto``):

```
[==========] Running 8 tests from 1 test case.           
[----------] Global test environment set-up.             
[----------] 8 tests from AES                            
[ RUN      ] AES.ValidateLegacy                          
[       OK ] AES.ValidateLegacy (1 ms)                   
[ RUN      ] AES.ValidateSecret                          
[       OK ] AES.ValidateSecret (0 ms)                   
[ RUN      ] AES.Encrypt                                 
[       OK ] AES.Encrypt (0 ms)                          
[ RUN      ] AES.EncryptNoBl                             
[       OK ] AES.EncryptNoBl (0 ms)                      
[ RUN      ] AES.Decrypt                                 
[       OK ] AES.Decrypt (0 ms)                          
[ RUN      ] AES.DecryptNoBl                             
[       OK ] AES.DecryptNoBl (0 ms)                      
[ RUN      ] AES.Loop                                    
[       OK ] AES.Loop (50 ms)                            
[ RUN      ] AES.LoopKey                                 
100000 encoded in 0.099907                               
[       OK ] AES.LoopKey (100 ms)                        
[----------] 8 tests from AES (151 ms total)             
                                                         
[----------] Global test environment tear-down           
[==========] 8 tests from 1 test case ran. (151 ms total)
[  PASSED  ] 8 tests.      
```

Signed-off-by: Piotr Dałek <piotr.dalek@corp.ovh.com>