@branch-predictor 2018-07-27
@rzarzynski As stated in commit comment and PR description, EVP API provides encapsulated access to specialised hardware (again - be it AES-NI, Padlock, or whatever world came up with). You can verify it with ``openssl speed``:
```
[branch@predictor ~]$ openssl speed aes-128-cbc
Doing aes-128 cbc for 3s on 16 size blocks: 14378122 aes-128 cbc's in 2.97s
Doing aes-128 cbc for 3s on 64 size blocks: 3973446 aes-128 cbc's in 2.97s
Doing aes-128 cbc for 3s on 256 size blocks: 1025526 aes-128 cbc's in 2.98s
Doing aes-128 cbc for 3s on 1024 size blocks: 261927 aes-128 cbc's in 2.98s
Doing aes-128 cbc for 3s on 8192 size blocks: 32869 aes-128 cbc's in 2.97s
OpenSSL 1.0.1e-fips 11 Feb 2013
built on: Wed Mar 22 21:37:40 UTC 2017
options:bn(64,32) md2(int) rc4(8x,mmx) des(ptr,risc1,16,long) aes(partial) idea(int) blowfish(idx)
compiler: gcc -fPIC -DOPENSSL_PIC -DZLIB -DOPENSSL_THREADS -D_REENTRANT -DDSO_DLFCN -DHAVE_DLFCN_H -DKRB5_MIT -DL_ENDIAN -DTERMIO -Wall -O2 -g -pipe -Wall -Wp,-D_FORTIFY_SOURCE=2 -fexceptions -fstack-protector --param=ssp-buffer-size=4 -m32 -march=i686 -mtune=atom -fasynchronous-unwind-tables -Wa,--noexecstack -DPURIFY -DOPENSSL_BN_ASM_PART_WORDS -DOPENSSL_IA32_SSE2 -DOPENSSL_BN_ASM_MONT -DOPENSSL_BN_ASM_GF2m -DSHA1_ASM -DSHA256_ASM -DSHA512_ASM -DMD5_ASM -DRMD160_ASM -DAES_ASM -DVPAES_ASM -DWHIRLPOOL_ASM -DGHASH_ASM
The 'numbers' are in 1000s of bytes per second processed.
type             16 bytes     64 bytes    256 bytes   1024 bytes   8192 bytes
aes-128 cbc      77457.90k    85623.08k    88098.88k    90004.45k    90660.89k
```

```
[branch@predictor ~]$ openssl speed -evp aes-128-cbc
Doing aes-128-cbc for 3s on 16 size blocks: 89705436 aes-128-cbc's in 2.96s
Doing aes-128-cbc for 3s on 64 size blocks: 21338855 aes-128-cbc's in 2.97s
Doing aes-128-cbc for 3s on 256 size blocks: 5967082 aes-128-cbc's in 2.97s
Doing aes-128-cbc for 3s on 1024 size blocks: 1540793 aes-128-cbc's in 2.97s
Doing aes-128-cbc for 3s on 8192 size blocks: 193836 aes-128-cbc's in 2.96s
OpenSSL 1.0.1e-fips 11 Feb 2013
built on: Wed Mar 22 21:37:40 UTC 2017
options:bn(64,32) md2(int) rc4(8x,mmx) des(ptr,risc1,16,long) aes(partial) idea(int) blowfish(idx)
compiler: gcc -fPIC -DOPENSSL_PIC -DZLIB -DOPENSSL_THREADS -D_REENTRANT -DDSO_DLFCN -DHAVE_DLFCN_H -DKRB5_MIT -DL_ENDIAN -DTERMIO -Wall -O2 -g -pipe -Wall -Wp,-D_FORTIFY_SOURCE=2 -fexceptions -fstack-protector --param=ssp-buffer-size=4 -m32 -march=i686 -mtune=atom -fasynchronous-unwind-tables -Wa,--noexecstack -DPURIFY -DOPENSSL_BN_ASM_PART_WORDS -DOPENSSL_IA32_SSE2 -DOPENSSL_BN_ASM_MONT -DOPENSSL_BN_ASM_GF2m -DSHA1_ASM -DSHA256_ASM -DSHA512_ASM -DMD5_ASM -DRMD160_ASM -DAES_ASM -DVPAES_ASM -DWHIRLPOOL_ASM -DGHASH_ASM
The 'numbers' are in 1000s of bytes per second processed.
type             16 bytes     64 bytes    256 bytes   1024 bytes   8192 bytes
aes-128-cbc     484894.25k   459827.18k   514334.34k   531236.37k   536454.23k
```

EVP (AES-NI) implementation was over 6x faster on **16-byte** blocks -- and this is on my low-end VPS that happens to have AES-NI enabled. I've seen similar results on the bare metal Xeon boxes.

By using ``AES_cbc_encrypt`` directly, you're forcing usage of "highly optimized, Perl-crafted assembly" version that can't make use of any crypto accelerators available, including AES-NI. Depending on package maintainer, ``AES_cbc_encrypt`` may even not support SSE2, degrading performance even further.